fix: does not store null params

+++ src/routes/landing/+page.ts
@@ -11,24 +11,25 @@ export function load(loadEvent) {
 	const study_id: string | null = params.get('study_id');
 	const session_id: string | null = params.get('session_id');
 
-	window.localStorage.clear();
-	window.localStorage.setItem(
-		'prolific_params',
-		JSON.stringify({ prolific_pid, study_id, session_id })
-	);
+	if (prolific_pid !== null && study_id !== null && session_id !== null) {
+		window.localStorage.clear();
+		window.localStorage.setItem(
+			'prolific_params',
+			JSON.stringify({ prolific_pid, study_id, session_id })
+		);
 
-	if (!window.chrome || !window.chrome.runtime) {
-		goto('../installation');
+		if (!window.chrome || !window.chrome.runtime) {
+			goto('../installation');
+		} else {
+			chrome.runtime.sendMessage(PUBLIC_EXTENSION_ID, {
+				type: 'is_installed'
+			});
+		}
 	}
-
-	const isExtensionInstalled: string | null = chrome.runtime.sendMessage(PUBLIC_EXTENSION_ID, {
-		type: 'is_installed'
-	});
 
 	return {
 		prolific_pid,
 		study_id,
-		session_id,
-		isExtensionInstalled // is this used?
+		session_id
 	};
 }
